+++ docs/source/changelog.rst
@@ -7,6 +7,10 @@ All notable changes to this project will be documented in this file.
 
 `Unreleased`_
 -------------
+
+`0.4.0`_ -- 2018-11-15
+----------------------
+
 Changed
 ^^^^^^^
 
@@ -70,6 +74,8 @@ Fixed
 
 * pipe_ignore was not recovered from right message
 
-.. _Unreleased: https://github.com/wiremind/remoulade/compare/v0.2.0...HEAD
+.. _Unreleased: https://github.com/wiremind/remoulade/compare/v0.4.0...HEAD
+.. _0.4.0: https://github.com/wiremind/remoulade/releases/tag/v0.4.0
+.. _0.3.0: https://github.com/wiremind/remoulade/releases/tag/v0.3.0
 .. _0.2.0: https://github.com/wiremind/remoulade/releases/tag/v0.2.0
 .. _0.1.0: https://github.com/wiremind/remoulade/releases/tag/v0.1.0